+++ CoustomerServices/src/main/java/com/example/coustomerservices/service/CustomerService.java
@@ -6,6 +6,7 @@ import com.example.coustomerservices.dto.NotificationDTO;
 import com.example.coustomerservices.entity.Customer;
 import com.example.coustomerservices.feignconfig.NotificationServiceClient;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -20,16 +21,24 @@ public class CustomerService {
     @Autowired
     private NotificationServiceClient notificationServiceClient;
 
+    //all customers
     public List<CustomerDTO> getAllCustomers() {
         return customerRepository.findAll().stream().map(this::convertToDTO).collect(Collectors.toList());
     }
 
+    //customer by id
     public CustomerDTO getCustomerById(Long id) {
         Customer customer = customerRepository.findById(id).orElse(null);
         return customer != null ? convertToDTO(customer) : null;
     }
 
-    public void createCustomer(CustomerDTO customerDTO) {
+    //customer creation
+    public String  createCustomer(CustomerDTO customerDTO) {
+
+        //customer already exists
+        if (customerRepository.existsCustomerByEmail(customerDTO.getEmail())) {
+            return "Customer with email " + customerDTO.getEmail() + " already exists use different email.";
+        }
 
         // Create and save the new customer
         Customer customer = new Customer();
@@ -44,10 +53,16 @@ public class CustomerService {
         notificationServiceClient.sendNotification(notificationDTO);
 
         customerRepository.save(customer);
+        return "Customer created successfully";
     }
 
-    public void deleteCustomer(Long id) {
+    //customer deletion
+    public String deleteCustomer(Long id) {
+        if(!customerRepository.existsById(id)) {
+            return "Customer with id " + id + " does not exist.";
+        }
         customerRepository.deleteById(id);
+        return "Customer deleted successfully";
     }
 
     private CustomerDTO convertToDTO(Customer customer) {
